added authorization

+++ src/tasks/tasks.service.ts
@@ -8,6 +8,7 @@ import { Task } from './task.entity';
 import { CreateTaskDto } from './create-task.dto';
 import { TaskStatus } from './task-status.enum';
 import { GetTasksFilterDto } from './get-tasks-filter.dto';
+import { User } from '../auth/user.entity';
 
 @Injectable()
 export class TasksService {
@@ -19,9 +20,12 @@ constructor(
 
 
 
-async getTasks(filterDto:GetTasksFilterDto):Promise<Task[]>{
+async getTasks(filterDto:GetTasksFilterDto,
+  user: User
+  
+  ):Promise<Task[]>{
 
-  return this.taskRepository.getTasks(filterDto)
+  return this.taskRepository.getTasks(filterDto,user)
 
 }
 
@@ -32,9 +36,15 @@ async getTasks(filterDto:GetTasksFilterDto):Promise<Task[]>{
 
 
 
-async getTaskById(id: number) :Promise<Task> {
+async getTaskById(id: number,
+  user: User
 
-  const found= await this.taskRepository.findOne(id)
+  
+  ) :Promise<Task> {
+
+  const found= await this.taskRepository.findOne({ 
+    where:{id,userId:user.id}
+  })
 
   if(!found){
  throw new NotFoundException(`Task with ID ${id} not found`)
@@ -45,30 +55,35 @@ return found
 }
 
 
-async createTask(createTaskDto:CreateTaskDto): Promise<Task> {
+async createTask(
+  createTaskDto:CreateTaskDto,
+  user:User
+  ): Promise<Task> {
 
-  return this.taskRepository.createTask(createTaskDto)
+  return this.taskRepository.createTask(createTaskDto,user)
 
 
 }
 
 
-async deleteTask(id:number):Promise<void>{
+async deleteTask(id:number,user:User):Promise<void>{
 
-const result=await this.taskRepository.delete(id)
+const result=await this.taskRepository.delete({id,userId:user.id})
+// where clause is not used here
 
 if(result.affected === 0){
   throw new NotFoundException(`Task with ID ${id} not found`)
 }
 
-// console.log(result)
 
 }
 
 
-async updateTaskStatus(id: number,status:TaskStatus):Promise <Task> {
+async updateTaskStatus(id: number,status:TaskStatus,
+  user: User
+  ):Promise <Task> {
 
-  const task=await  this.getTaskById(id)
+  const task=await  this.getTaskById(id,user)
 
 task.status=status
